+++ PowerShellScripts/UpdateAndPack/CoreProAutoUpdateAndPack.ps1
@@ -1,6 +1,7 @@
 Param(
     [string]$ProjectDir,
     [String]$GitWorker,
+    [String]$GitMainBranch = "master",
     [String]$NugetServerPath
 )
 
@@ -18,20 +19,23 @@ if ([string]::IsNullOrEmpty($GitWorker)) {
 Write-Host "==========" + $ProjectDir + "==========="
 Set-Location $ProjectDir
 
-Write-Host "Step1. Check csproj" -ForegroundColor Green
+Write-Host "===>Step1. Check csproj" -ForegroundColor Green
 if ((Get-ChildItem '*.csproj').Length -eq 0) {
     Write-Error ".csproj not found"
     return
 }
 
-Write-Host "Step2. Git Pull" -ForegroundColor Green
+Write-Host "===>Step2.1. Git Switch Branch | Command: git checkout $GitMainBranch" -ForegroundColor Green
+git checkout $GitMainBranch
+
+Write-Host "===>Step2. Git Pull | Command: git pull" -ForegroundColor Green
 bash $GitWorker
 
 $BuildCommand = "dotnet build"
-Write-Host "Step3. Build | Command: $BuildCommand" -ForegroundColor Green
+Write-Host "===>Step3. Build | Command: $BuildCommand" -ForegroundColor Green
 Invoke-Expression $BuildCommand
 
-$PackCommand = "dotnet pack"
+$PackCommand = "===>dotnet pack"
 Write-Host "Step4. Pack | Command: $PackCommand" -ForegroundColor Green
 Invoke-Expression $PackCommand
 
@@ -44,12 +48,11 @@ if ($PackResult.Length -gt 0) {
     if ([string]::IsNullOrEmpty($NugetServerPath)) {
         Write-Warning "===>Nupkg has success generated on $PackageFileObj, but [NugetServerPath] is NullOrEmpty."
     } else {
-        Write-Host "${NugetServerPath}\$($PackageFileObj.Name)"
-        if (Test-Path "${NugetServerPath}\$($PackageFileObj.Name)") {
-            Write-Warning "Package Exited"
+        if (!$NugetServerPath.StartsWith("http") && Test-Path "${NugetServerPath}\$($PackageFileObj.Name)") {
+            Write-Warning "Package $PackageFileObj Exited"
         } else {
             $PushCommand = "dotnet nuget push ${PackageFileObj} -s $NugetServerPath --skip-duplicate"
-            Write-Host "Step5. Push | Command: $PushCommand" -ForegroundColor Green
+            Write-Host "===>Step5. Push | Command: $PushCommand" -ForegroundColor Green
             Invoke-Expression $PushCommand
         }
     }
